+++ src/main/java/io/github/xf8b/morefeatures/enchantments/HarmingCurse.java
@@ -1,8 +1,8 @@
 package io.github.xf8b.morefeatures.enchantments;
 
+import io.github.xf8b.morefeatures.config.MoreFeaturesConfig;
 import io.github.xf8b.morefeatures.core.MoreFeatures;
 import io.github.xf8b.morefeatures.core.MoreFeaturesRegistries;
-import io.github.xf8b.morefeatures.config.MoreFeaturesConfig;
 import io.github.xf8b.morefeatures.util.handler.TickHandler;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentHelper;