+++ src/main/java/io/github/xf8b/morefeatures/items/armor/corrupted/CorruptedLeggings.java
@@ -2,9 +2,9 @@ package io.github.xf8b.morefeatures.items.armor.corrupted;
 
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
+import io.github.xf8b.morefeatures.config.MoreFeaturesConfig;
 import io.github.xf8b.morefeatures.core.MoreFeatures;
 import io.github.xf8b.morefeatures.items.armor.MoreFeaturesArmorMaterial;
-import io.github.xf8b.morefeatures.config.MoreFeaturesConfig;
 import io.github.xf8b.morefeatures.util.helpers.KeyboardChecker;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.entity.SharedMonsterAttributes;
@@ -14,7 +14,7 @@ import net.minecraft.item.ArmorItem;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.StringTextComponent;
+import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraft.world.World;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
@@ -42,11 +42,11 @@ public class CorruptedLeggings extends ArmorItem {
     @Override
     public void addInformation(ItemStack stack, World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
         if (KeyboardChecker.isHoldingShift()) {
-            tooltip.add(new StringTextComponent("This armor has been corrupted."));
-            tooltip.add(new StringTextComponent("The protection and toughness will be random."));
-            tooltip.add(new StringTextComponent("Good luck, and pray to the RNG gods."));
+            tooltip.add(new TranslationTextComponent("tooltip." + MoreFeatures.MOD_ID + ".corrupted_armor.line.one"));
+            tooltip.add(new TranslationTextComponent("tooltip." + MoreFeatures.MOD_ID + ".corrupted_armor.line.two"));
+            tooltip.add(new TranslationTextComponent("tooltip." + MoreFeatures.MOD_ID + ".corrupted_armor.line.three"));
         } else {
-            tooltip.add(new StringTextComponent("Hold " + "\u00A7e" + "SHIFT " + "\u00A7r" + "to see info."));
+            tooltip.add(new TranslationTextComponent("tooltip." + MoreFeatures.MOD_ID + ".hold_shift"));
         }
         super.addInformation(stack, worldIn, tooltip, flagIn);
     }